requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Pillow==9.5.0
SQLAlchemy==2.0.15
typing_extensions==4.6.1
Werkzeug==2.3.4
